Add options to disable segmentation parameters by using negative values

+++ scripts/mesh_object.py
@@ -24,6 +24,13 @@ from sensor_msgs.msg import (
   PointCloud2
 )
 
+"""
+Examples:
+rosrun object_reconstruction mesh_object.py -i right_frame.bag --leaf_size 0.002 --seg_z_min 0.01 --final_stat_filter_k 20 --final_stat_filter_std_dev 1 --final_leaf_size 0.001 --stat_filter_k -1 --seg_radius_crop -1
+
+rosrun object_reconstruction mesh_object.py -i left_frame.bag --leaf_size 0.002 --seg_z_min 0.01 --final_stat_filter_k 50 --final_stat_filter_std_dev 1 --final_leaf_size 0.001 --stat_filter_k -1 --seg_radius_crop -1
+"""
+
 
 FILTER_SCRIPT_PIVOTING="""
 <!DOCTYPE FilterScript>
@@ -43,6 +50,7 @@ FILTER_SCRIPT_PIVOTING="""
 </FilterScript>
 """
 
+logger = criros.utils.TextColors()
 
 class MeshObject(object):
   def __init__(self, options):
@@ -58,8 +66,8 @@ class MeshObject(object):
     self.final_stat_filter_k = options.final_stat_filter_k
     self.final_stat_filter_std_dev = options.final_stat_filter_std_dev
     self.final_leaf_size = options.final_leaf_size
+    self.stl = options.stl
     self.basename = os.path.splitext(os.path.abspath(options.bag))[0]
-    rospy.on_shutdown(self.on_shutdown)
   
   def consistent_bag(self):
     # check we have the same number of msgs
@@ -70,7 +78,7 @@ class MeshObject(object):
   def execute(self):
     consistent, num_msgs = self.consistent_bag()
     if not consistent:
-      rospy.logerr( 'The input bag is not consistent.' 
+      logger.logerr( 'The input bag is not consistent.' 
                     'Check that all the topics have the same number of messages.')
       return
     # Get the pattern poses and estimate the table's center
@@ -103,12 +111,16 @@ class MeshObject(object):
       inliers = []
       for i,point in enumerate(cloud_xyz):
         inside_z_range = self.seg_z_min < plane.distance(point) < self.seg_z_crop
-        xy_dist = tr.vector_norm(point[:2]-table_center[:2])
-        if (inside_z_range and xy_dist < self.seg_radius_crop):
+        if self.seg_radius_crop > 0:
+          inside_radius = tr.vector_norm(point[:2]-table_center[:2]) < self.seg_radius_crop
+        else:
+          # A negative value disables this radius cropping
+          inside_radius = True
+        if (inside_z_range and inside_radius):
           inliers.append(i)
       cloud = pcl.PointCloud( cloud_xyz[inliers] )
       # Statistical outlier filter x 2
-      for i in range(2):
+      if self.stat_filter_k > 0:
         stat_filter = cloud.make_statistical_outlier_filter()
         stat_filter.set_mean_k(self.stat_filter_k)
         stat_filter.set_std_dev_mul_thresh(self.stat_filter_std_dev)
@@ -124,17 +136,19 @@ class MeshObject(object):
       pbar.update(num_msgs-len(poses))
     self.bag.close()
     pbar.finish()
-    # Final statistical outlier filter
     cloud = pcl.PointCloud( all_points )
-    stat_filter = cloud.make_statistical_outlier_filter()
-    stat_filter.set_mean_k(self.final_stat_filter_k)
-    stat_filter.set_std_dev_mul_thresh(self.final_stat_filter_std_dev)
-    cloud = stat_filter.filter()
+    # Final statistical outlier filter
+    if self.final_stat_filter_k > 0:
+      stat_filter = cloud.make_statistical_outlier_filter()
+      stat_filter.set_mean_k(self.final_stat_filter_k)
+      stat_filter.set_std_dev_mul_thresh(self.final_stat_filter_std_dev)
+      cloud = stat_filter.filter()
     # Final voxel grid filter
-    voxel_filter = cloud.make_voxel_grid_filter()
-    leaf_size = np.ones(3)*self.final_leaf_size
-    voxel_filter.set_leaf_size(*leaf_size)
-    cloud = voxel_filter.filter()
+    if self.final_leaf_size > 0:
+      voxel_filter = cloud.make_voxel_grid_filter()
+      leaf_size = np.ones(3)*self.final_leaf_size
+      voxel_filter.set_leaf_size(*leaf_size)
+      cloud = voxel_filter.filter()
     # Generate files
     devnull = open('/dev/null', 'w')
     pcd_filename = self.basename + '.pcd'
@@ -142,27 +156,24 @@ class MeshObject(object):
     stl_filename = self.basename + '.stl'
     # PCD
     cloud.to_file(pcd_filename)
-    rospy.loginfo('Saved PCD file: %s' % pcd_filename)
+    logger.loginfo('Resulting cloud has {0} points'.format(cloud.size))
+    logger.loginfo('Saved PCD file: %s' % pcd_filename)
     # PLY
     env = os.environ
-    #  object_mesh.pcd object_mesh.ply
     subprocess.Popen('pcl_pcd2ply -format 1 -use_camera 0 ' + pcd_filename + ' ' + ply_filename,
                      env=env, cwd=os.getcwd(), shell=True, stdout=devnull).wait()
-    rospy.loginfo('Saved PLY file: %s' % ply_filename)
+    logger.loginfo('Saved PLY file: %s' % ply_filename)
     # STL
-    f = tempfile.NamedTemporaryFile(delete=False)
-    script = f.name
-    f.write(FILTER_SCRIPT_PIVOTING)
-    f.close()
-    env['LC_NUMERIC'] = 'C'
-    subprocess.Popen('meshlabserver -i ' + ply_filename + ' -o ' + stl_filename + ' -s ' + script,
-                     env=env, cwd=os.getcwd(), shell=True, stdout=devnull, stderr=devnull).wait()
-    os.remove(script)
-    rospy.loginfo('Saved STL file: %s' % stl_filename)
-  
-  def on_shutdown(self):
-    # Close rosbag
-    self.bag.close()
+    if self.stl:
+      f = tempfile.NamedTemporaryFile(delete=False)
+      script = f.name
+      f.write(FILTER_SCRIPT_PIVOTING)
+      f.close()
+      env['LC_NUMERIC'] = 'C'
+      subprocess.Popen('meshlabserver -i ' + ply_filename + ' -o ' + stl_filename + ' -s ' + script,
+                       env=env, cwd=os.getcwd(), shell=True, stdout=devnull, stderr=devnull).wait()
+      os.remove(script)
+      logger.loginfo('Saved STL file: %s' % stl_filename)
 
 
 def parse_args():
@@ -199,19 +210,20 @@ def parse_args():
   parser.add_argument('--final_leaf_size', metavar='FINAL_LEAF_SIZE', type=float, default= 0.002,
                       help='Leaf size (meters) to be used to downsample + filter' 
                       'the initial point cloud. default=%(default).3f')
+  parser.add_argument('--stl', action='store_true',
+                      help='If set, will generate a STL file from the point cloud')
   parser.add_argument('--debug', action='store_true',
                       help='If set, will show additional debugging information')
   args = parser.parse_args(rospy.myargv()[1:])
   if len(args.bag) < 1:
     parser.print_help()
-    print '\nYou must supply the input bag name'
+    logger.logerr('\nYou must supply the input bag name')
     sys.exit(1)
   return args
 
 if "__main__" == __name__:
   options = parse_args()
   log_level= rospy.DEBUG if options.debug else rospy.INFO
-  node_name = os.path.splitext(os.path.basename(__file__))[0]
-  rospy.init_node(node_name, log_level=log_level)
+  logger.set_log_level(log_level)
   meshit = MeshObject(options)
   meshit.execute()
